server/requirements.txt

Flask
mysql-connector-python
bcrypt
python-dotenv
